Create core.magics.logging according to new API.
