+++ IPython/core/magics/__init__.py
@@ -12,13 +12,14 @@
 # Imports
 #-----------------------------------------------------------------------------
 
-from IPython.core.magic import Magics, register_magics
+from ..magic import Magics, register_magics
 from .auto import AutoMagics
 from .basic import BasicMagics
 from .code import CodeMagics, MacroToEdit
 from .config import ConfigMagics
 from .execution import ExecutionMagics
 from .history import HistoryMagics
+from .logging import LoggingMagics
 from .namespace import NamespaceMagics
 from .osm import OSMagics
